fix out-of-memory error when generating large GPX
In some cases we managed to generate GPX faster than it could be sent
over the network leading to out-of-memory issues related to buffer
exhaustion.
A proper solution for such issues is to convert our generation process
to a ReadableStream that can be piped to Response Stream (back pressure
mechanism will prevent us from generating text too quickly).

Turns out the simplest way of converting export code to Readable stream
is to use [generators][1] (yielding in the middle of the function
preserves the process 'stack'). Converting a generator to a readable
stream is quite trivial - we use [stream-generators][2] module for that.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/function*
[2]: https://www.npmjs.com/package/stream-generators